Add a job to send multiple rows to Geckoboard
`GeckoboardEvent` currently sends data to Geckoboard for one claim, one at
a time. This is fine for the majority of purposes, but if we want to submit a
load of data to Geckoboard, we'll be rate limited (https://developer.geckoboard.com/hc/en-us/articles/360019949831).

This adds a class, which extends from the `GeckoboardEvent` class, which
allows us to send an array of rows to Geckoboard, so the backfill will be
carried out in one API call.