Wire in the post-eligiblity Qs for Maths & Physics
This will get the user as far as the check-your-answers page. For now,
the check-your-answers page is just a holding page as the real thing
is planned to be implemented as a separate story.

This commit adds a temporary check-your-answers template as the existing
one will break if given a Maths & Physics claim. A separate story is
planned that will make check-your-answers work with any type of claim.

Note that there is now a fair amount of commonality/duplication between
the two SlugSequence classes but it feels too early to try and extract
any kind of abstraction.