Update approved/rejected email content
Updated based on feedback from Faye. As well as updating the content to
be more helpful to the user, we're also now using a consistent sentence
when referring to the claim, which will make it easier to adjust these
emails based on the policy. We've also removed mention of the claim
school from the emails as this is not really necessary, and also
confusing for users that may have worked at multiple schools.