Clear claim session before prepending view path
Currently, if a user starts a claim session for one policy, and then attempts to
switch policies, they're confronted with a Template error. This is because
`prepend_view_path_for_policy` runs before the claim session is cleared, so
the view path is set for the old policy. This tweak changes this behaviour so
`clear_claim_session` runs before the view path is set.