Add the gross pay for payments
We've been using the gross value from the Cantium confirmation report
in our payment advice to the claimants.

This is incorrect, the 'gross value' as they call it is infact the total
outlay (how much it is costing DfE). Instead we need to subtract the
employer's (DfE) NI contributions from this value to get the gross pay
that will be communicated to the claimant.

I'm not sure that having both `gross_pay` and `gross_value` is a good
idea. If either of them should change it should be `gross_value` but
since that's the name that Cantium have given it I've decided to leave
it for now.

I decided against having `gross_pay` as a computed method as we will
probably want to query against it in the future.