Set cache headers so pages are refreshed each time
This fixes two issues - the first being an issue where a user who has finished,
or was part way through their claim and has navigated away, can have their
personal data surfaced by someone using the back button, even after the
session has expired.

The second issue is if a user starts the Verify journey and then uses the back
button, clicking continues results in an error, because the browser sends the
same authentication request, rather than a newly minted one. This is slightly
harder to test, because we stub out a lot of the Verify stuff in feature tests
(and we  don't even see the error in the development version of the VSP), so
the only thing we can test is that the cache headers get successfully set in
the part of the journey where the authentication request is made.