De-couple Payroll::ClaimCsvRow from Student Loans
This means Payroll::ClaimCsvRow won't fall over when presented with a
claim that isn't one for the Student Loans policy.

The award_amount method is the definitive and policy-agnostic method for
determining the amount for the payment. Although there is a
Claim#award_amount method I've chosen to go with the
Payment#award_amount method for the payroll report. Technically they
should always be the same as one gets set from the other when the
payment is created, but it feels more appropriate to treat the value on
the Payment as the final arbiter of the amount we send to Cantium for
grossing up.