Extract eligibility_answers helper into a class
We want the ClaimsHelper#eligibility_answers method for displaying a
claim's eligibility answers on the check-your-answers page to be
policy-agnostic. Our existing pattern for slotting in policy-specific
behaviour is to introduce a new class nested under the policy, so we do
the same here, introducing an EligibilityAnswersPresenter class which
the policy provides in order to format the eligibility answers.

The subject_list method which was previously in ClaimsHelper has been
moved to a new StudentLoans::PresenterMethods module so that we can
continue to use it in Admin::ClaimsHelper.