Make headless Chrome screen size bigger
We're having some issues with tests failing on Headless Chrome occasionally
when forms have hints that make the screen size a specific height. Working
theory is that Chromedriver isn't taking account of the margin/padding of the
span element and causing it to click in the wrong place. An easy fix for this is
to increase the screen size to make sure Chromedriver doesn't have to scroll
for most cases.