Introduce DfE Sign-in role for payroll operators
We'll use this to grant restricted access to Cantium employees so that
they can sign-in and download the monthly payroll file.

This new role has been set up on pre-production and is scoped to a new
"Cantium Business Solutions" organisation. That way, the approver for
this organisation will be able to grant and revoke access to users
themselves without the need for intervention from us. For now, I have
been set up as the approver on pre-production so that we can test the
behaviour whilst we build out the new functionality.