Add ITT subject specialism to Maths & Physics
This is second part of the revised flow for understanding a claimant's
eligiblity based on their qualifications. For now we're just adding the
new questions and not modifying the eligiblity code, or removing the
existing question. That will come in later commits.

The _prefix on the enums has become necessary because both ITT-related
enums share an answer (:physics), which results in a conflict in the
default boolean methods that ActiveRecord tries to define for all enum
options.