Move loan amount answer in “Check your answers”
Technically, it is easiest if the eligibility section is the only
policy-specific section on the “Check your answers” page. In student
loans, this is _almost_ the way things are, except for the student loan
amount question.

We thought about ways to keep things simple technically whilst not
affecting the user experience too much. We decided to not move the
question in the journey, but we are happy for the answer to be put
alongside the other eligibility answers on the “Check your answers”
page. It means that this question will not appear in the exact order in
which it was asked, but we think it’s an acceptable compromise.