Pass claims' policies to payroll runs factory
This allows us to use the payroll run factory to create a payroll run
with claims from multiple policies. I am making this change because I
want to update the local development seeds to include some Maths and
Physics claims.

I had to change the way that we create the claim association in
factories/payments.rb, putting it inside a block so that I could access
the transient claim_policy attribute. I found this out here:
https://hschne.at/2019/09/22/fabulous-factorybot.html