Auto click link to download the payroll run file
Using JavaScript, any link with a 'data-auto-follow-link' set to 'true' will
automatically click itself after 3 seconds.

Because the Payroll operator only gets one chance to download the payroll run
file we use this feature to start the download for them.

There is still the fallback of the user clicking the link manually, which is
made clear in the content on the page.