Copy the payroll run download url with javascript
A text field input with the 'data-copy-to-clipboard' set to 'true' will get a
'Copy to clipboard' button added, that when clicked, copies the content of the
field to the clipboard.

We used the older and simpler method of select() and execCommand('copy') as it
has the widest browser support and this is just a temporary solution until this
process is fully automated.