Update check-your-answers feature specs for M&P
This does three things:

- Slims down the check-your-answers specs, reducing it to a
representative set of scenarios instead of a separate scenario for
changing each answer from each policy, which would not be a sustainable
approach as we add new policies.

- Introduces some maths and physics questions into the
check-your-answers specs (as part of the aforementioned set of
scenarios).

- Rearranges the file so that it's clear which specs are related to
eligibility (and hence are only representative of a broader set of
policy-specific questions), and which specs represent questions that are
present across all policies.