Don't include ineligible slug when not appropriate
Instead of a raising a 500 error if a user visits the ineligible screen
with an claim that isn't actually ineligible we raise a 404.

We had an exception raised for the ineligible screen because a partial
for the ineligibility reason couldn't be found: "Missing partial
claims/_ineligibility_reason_". Based on the code in the partial, this
is likely to have happened because the claim eligibility didn't report
an ineligibility_reason. which would only happen if the claim wasn't
actually ineligible. This is probably a weird edge case and caused by
the user navigating outside of the normal UI interface, but in any case
Returning a 404 feels more appropriate and will reduce the rollbar
noise.

If we find more users ending up on the ineligible screen
unexpectedly we might want to consider handling things differently and
helping the user get back to their in-progress claim.