Clear JS timeout redirect when session refreshed
This fixes a bug where a user that has dismissed the timeout dialog to
continue their session still gets redirected after the timeout dialog
warning period has expired. This was happening because the JS timeout
that gets initially set to redirect the user after the timeout period
has expired wasn't being cleared.

It's unlikely many users will have experienced this because the timeout
is naturally cleared when a new page is loaded.