Configure implicit ordering based on created_at
One of the downsides to using UUIDs for primary keys is that we lose the
implicit ordering behaviour of the `first` and `last` finder methods,
which by default return the first and last records based on the primary
key as a proxy for creation date. When that is an incrementing integer,
things work as expected, but with UUIDs we end up with entirely
random records based on the string ordering of the UUIDs. Rails 6
introduced the ability to override this implicit order column[1], which
means we can set it to created_at and get back the natural and less
surprising behaviour of `first` and `last`.

[1]: https://github.com/rails/rails/pull/34480