Change Payment to Claim association to has_many
This is groundwork for being able to pay a person for multiple claims
with a single payment. We change the payment’s association with claims
to a "has one" instead of a "belongs to". This leaves us with a database
schema that will then allow us to change this association to a "has
many". This commit does not make any changes in business logic;
the app still assumes that a payment will only have one claim.

We do not yet remove the claim_id column from `payments`; this will come
in a future release.

Now that a payroll run's relation to its claims is a has_many through:,
the payroll_run#claims association does not work as (we) expected for a
new record [1]. So we have to change the PayrollRunContoller#new action
to directly pass claims and award amount to the view, instead of a
PayrollRun instance.

[1]: https://stackoverflow.com/questions/26651811/rails-activerecord-has-many-through-association-on-unsaved-objects

Co-authored-by: Robbie Paul <robbiepaul@me.com>