Payroll run files can be downloaded
Now we can record the download happening, this commit impliments the actual
download process.

We ask the Payroll opertator to initiate the download be clicking a button which
records the action in the database.

As the payroll run file contains large amounts of personally identifiable
information the download also triggers a 30 second timeout, after which a
Payroll operator can no longer download the file.

The final step in the download process it to show the Payroll operator a page
that either

- shows a link to the download file, if inside the timeout
- shows when and who downloaded the file if the timeout was reached

We decided that once the timeout is reached the request for the file should
redirect to this final page so the payroll operator can understand why there is
no longer a file available.