Correct schema.rb and currently failing migration
Turns out our schema.rb file was not an accurate reflection of the
actual database schema: it was missing indexes for created_at and
updated_at on the payroll_runs table. This corrects the schema.rb file
and also updates the most recent migration, which is currently failing
to run because it tries to add an index that already exists.

The problem stems from this commit:
https://github.com/DFE-Digital/dfe-teachers-payment-service/commit/39ca7bfa30144ff3c2a0671de42b984ad55d322b

It added the payroll_runs table with `t.timestamps index: true`, which
would have added indexes for created_at and updated_at, but for whatever
reason those indexes were not reflected in the schema.rb changes. The
result of this was our local versions of the database were likely to get
out of sync after running db:reset or similar commands that load the
schema. Same with the null constraint on the created_by column. What
probably happened is the migration was edited after it was intially run
during local development but wasn't subsequently re-run to update the
schema.rb file accordingly.

Note: developers will want to rebuild and reseed their local databases
after this is merged to get everything back in sync and avoid similar
problems in the future:

    rails db:reset