Check the slot responds before swapping slots
This ensures the container has come up cleanly before swapping it to the
production container. This is only done if we have an endpoint to check, so
we still have to hit an hope for the VSP, as it's only accessible from a limited
range of IP addresses. We could fix this, so it can be accessed from Azure
DevOps, but given the amount of development work we actually do on the
VSP is pretty much nil, I think it's safe to assume it'll always come up cleanly.