Use statutory high age to check school eligibility
For both maths and physics and student loan policies, special schools should
only be eligible when their statutory_high_age is over 11 and they are not a
post 16 institution.

Special schools all have the phase 'not_applicable' so there is no need to check
that attribute as it is implied.