Alternative provision schools are eligible
Alternative provisons schools are those with `school_type` of:

  - pupil_referral_unit
  - secure_unit
  - free_school_alternative_provider
  - academy_alternative_provision_converter
  - academy_alternative_provision_sponsor_led

In order to be eligible they will also to be secondary equivalent (i.e teach
students over 11).

Although secure units are state funded, the Get information about schools data
always lists the school_type_group as 'other' which means we do not consider
them when checking if they are state funded.

We cannot just include 'other' in STATE_FUNDED_SCHOOL_TYPE_GROUPS as it includes
other schools that are not state funded.