Make payments CSV classes use payments, not claims
This is prep work for updating the CSV to work with multiple claims,
where each line of the CSV will correspond to a single _payment_, not
necessarily a single claim. We modify and rename the CSV classes to
reflect this.

When we move to multiple claims, there will be some data fields which
are duplicated across a payment's claims, for example address, payroll
gender, etc. We add these fields as methods on the Payment class. At the
moment, they just delegate to the payment's claim. Once we move to
multiple claims, they will probably delegate to the first claim of the
payment. This allows us to keep the CSV generation code the same, and
abstract away the process of getting e.g. an address for a payment.