Update payroll CSV to handle payments with multiple claims
We update the existing tests instead of adding new ones, since a payment
with one claim can be considered an unexceptional example of a payment
with one or more claims.

Confirmed with Cantium that SCHEME_NAME is not needed unless DfE want to
receive breakdowns of costings by policy, which they do not currently.
We have hence decided to remove the column.

Now that each row of the CSV represents a payment instead of a claim, we
replace the CLAIM_ID column with PAYMENT_ID.

Cantium have already updated their system to handle the changes in
SCHEME_NAME and CLAIM_ID.

Co-authored-by: Robbie Paul <robbie.paul@dxw.com>