Remove Payment#claim shim method
All of the code has now been updated to use the #claims association.

We pick an arbitrary claim to use for a payment's personal details,
under the assumption that all of the payment's claims' details will
agree (which will be enforced as part of the claim checking process in a
future story).

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>