Add Claim#identity_confirmed? predicate method
This will be used to indicate whether a claim has a confirmed identity
or not. Right now, this is only true if the claim has a GOV.UK Verify
response associated with it (i.e. it has completed the GOV.UK Verify
stage of the claim journey), but future work will add the ability to
confirm the identity on a claim via other means. We'll use this method
in the meantime to prevent non-confirmed claims from being approved.