Allow matching attributes to be grouped together
This changes the `ATTRIBUTES_TO_MATCH` array to be an array of groups
to match against in the database. If there is more than one attribute in the
group, we concatenate the subjects and predicates together, so we can
make sure both things match. At the moment, we only need this for bank
details, but we may need this for other things (such as maybe firstname /
lastname / DOB etc)