Refactor maths and physics school eligibility spec
The exisiting spec used a great deal of nested contexts, instead I've tried to
demonstate what makes a school eligible or not. For maths and physics there are
four attributes we use to determine eligibility:

- the school is open
- the school is state funded
- the school is in an eligble area
- the school teaches secondary or it's equivalent

I am hoping this is a simpler document of what makes a school eligble and also
gives us appropriate test coverage.