Include GOV.UK Verify attributes in Claim factory
These will be present on a user-submitted claim (at least until we add
the ability to submit a claim without going through verify) and therefor
we should include them on the factory.

Developers will probably want to re-seed their local database so that
the claims that the seeds.rb file creates includes these attributes.