Updated content for ineligibility snippet
We found claimants were contacting support to confirm eligibility
criteria which was over burdening support staff. By updating all the
ineligibility content we hope to make it simpler for claimants to
understand why they may not be eligible.

This snippet is used for both maths and physics and student loans
policies.

The policy specific content will be updated in future commits.