Unpin Bundler version in Dockerfile
We locked the version of Bundler to 2.0.2 in the Dockerfile because with
the latest version (2.1.0) the app was failing to start – the calls to
the `rails` command in docker-entrypoint were resulting in "command not
found" errors. We traced the issue to the fact that bundler 2.1.0 when
installed in the Docker container doesn't appear to be creating any
binstubs for commands like `rake` or `rails`. Instead of relying on the
presence of binstubs we call the commands with `bundle exec` instead.