Introduce page to capture claimant's date of birth
The scoped validations for the date_of_birth attribute were already
present on the model (leftover from before GOV.UK Verify integration was
completed and we were capturing all attributes directly from the user)
so we don't have to add them here.

This means it is now technically possible to complete and submit a claim
without completing GOV.UK Verify if you visit the appropriate URLs.