Find or create a user in the database at login
This finds or creates a user by the User ID that we get back from DfE Signin
at the point of login. We then store the internal ID (rather than the DfE Signin
ID) in the session, so we can access the user's details to record which user
has carried out an action.