Introduce endpoint for skipping GOV.UK Verify
This gives us a single, policy-agnostic endpoint that will take a user
to the appropriate page when they can't get through GOV.UK Verify. We
need a single endpoint as the page they will be linked from is static
content on GOV.UK VErify that is the same across all policies.