Add BUILD_NUMBER env var to containers
This adds a new environment variable to the web and worker containers, set
by setting the `AZURE_BUILD_NUMBER` in the build environment when
deploying. This means that the worker container knows that we have new
code to deploy and pulls down the latest version of the development or
production image from DockerHub.