Mailers extend from ApplicationMailer
The only mailers we use are `mail-notify` mailers so we can change our
(currently unused) ApplicationMailer to extend the Mail::Notify::Mailer
then extend our mailers from that.

The GOV.UK Notify template ID isn't a secret, nor is it different
between environments and it isn't likely to change so it should be safe
to hardcode it.