Creator of a PayrollRun is recorded as a reference
This adds a new reference to the Payroll Runs table, which is a reference to
a DfeSignIn::User, rather than a bare string. I'll need to keep the original
created_by column around to carry out the data migration for now, but in the
meantime I've had to remove the not null constraint from that column, as
the `created_by` methods now refer to the referenced model.