Set up the Azure test environment
Previously we've been testing out infrastructure changes by deploying
a tmp app to the `development` subscription. This was working well until
we needed to make changes to the virtual network where creating a tmp
network with it's own subnet will be difficult and probably cause
problems for the main app in development.

This commit introduces the templates necessary to deploy our
infrastructure to a 'test' subscription and not interfere with
development in any way.