Create a single payment for multiple claims from same claimant
If the same person claims for multiple policies and their claims are
approved in the same payroll month, then HMRC rules mean that we must
combine their two awards into a single payment.

We use National Insurance number to identify a claimant. A future commit
will introduce validation to ensure that (some subset of) the claimant's
personal details are consistent across the claims. We are also going to
talk to claims checkers to ensure that they will not approve a claim if
the claimant's National Insurance number does not match that found in
DQT or Hartlink.