Validate that personal details of a payment's claims are consistent
Given a payment with multiple claims, we want to ensure that there is a
single, unambiguous value for _some_ of the columns in the payroll file
which are derived from the payments' claims. The columns that we ensure
are consistent are those which either could lead to money being paid
into the wrong bank account, or which could affect the amount that the
claimant receives.

There are other personal details attributes which could legitimately
change over time, and for which we believe it will not be a problem if
we simply pick the latest value and send it to Cantium (and hence HMRC).
Hence we do not validate the consistency of these attributes. A future
commit will introduce the change to pick the value from the latest claim
instead of the earliest one, as it is at the moment.

Eventually, we will need to catch the forbidden inconsistencies much
earlier, as part of the claims checking journey
(https://trello.com/c/vqAzWiOZ/1114-flag-payroll-information-discrepancies-to-claim-checkers),
so that inconsistent claims can be caught before we get as far as
blocking the payroll process. Until that is implemented, these
validations prevent us from generating a payroll file based on
inconsistent data.

Co-authored-by: Robbie Paul <robbie.paul@dxw.com>