Explicitly mark West Somerset schools as eligible
The policy document says that schools in the West Somerset opportunity
area are eligible, but this is defined as contiguous with the (now
defunct) West Somerset local authority district. When the local
authority districts of West Somerset and Taunton were merged this local
authority district ceased to exist, and the new local authority district
which schools were moved to is not eligible.

Schools in the
`SCHOOL_URNS_CONSIDERED_AS_ELIGIBLE_LOCAL_AUTHORITY_DISTRICT` list are
now treated as being in an eligible local authority district. This is a
list of schools which are in the West Somerset opportunity area and
which are eligible.

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>