Reduce the chance of user submitting GOV.UK Verify SAML form twice
As we have observed in the past (e.g. commit 0f225f8), submitting the
same SAML request to Verify twice causes Verify to show a "Sorry,
something went wrong" error screen.

It was possible to submit the request twice by clicking the "Continue"
button on the `/verify/authentications/new` page after the JavaScript
auto-submit had been triggered. (I was able to reproduce this in a
browser, in our development environment, although not consistently.)

To prevent this, we follow an approach similar to that suggested by the
Verify documentation [1] and various Verify-using services which I
looked at, and we hide the submit button when JavaScript is enabled. We
also change the heading and explanatory messages so that the user knows
what's going on and how to recover from an error.

I've also updated the JS-disabled message to match other services, which
explain that the user needs to click the button _because_ JS is
disabled.

I'm not sure if there is anything we can do to prevent a double
submission in the non-JS case.

[1] https://www.docs.verify.service.gov.uk/get-started/set-up-successful-verification-journey/#send-the-authentication-request-to-the-testing-service