Fix database setup to load db schema and data schema files
When setting up a new environment we run "db:setup" which loads the db
schema and seeds the database. The next time the container starts it
attempts to run the data migrations, many of which fail to run.

Instead we should explicitly load the db and data schema files so no
needless migrations are ran on subsequent deploys.