User can continue claim after starting a new claim
This adds a button for a user to click to continue their claim. This
also means we need to store a user's place in the flow in the session
each time they hit the show action for a claim, which allows us to take
the user back to where they were in the flow.

Co-authored-by: Stuart Harrison <stuart@dxw.com>