Claim journey prefers the policy param from URL
Whilst the user is in a claim journey the policy which is named in the
URL is now preferred in `current_policy_routing_name`. If not present,
the method falls back to the policy of the current claim in the session.

This brings page titles, headers etc in line with the URL.