A user is warned about starting a new claim
When a user is mid-claim and tries to start a new claim by navigating to
the `new` action for either policy, they're warned that the progress
they have made will be lost (unless their claim is ineligible). They can
then click on "Start new application" to continue to make a new claim.

Co-authored-by: Stuart Harrison <stuart@dxw.com>