Ensure container instance has restarted before fetching the IP
It's possible that the container instance could be in a "Waiting" state
if it ran in to an error. This would break our script to fetch the IP
address as you can only execute commands on a running container
instance.

By restarting the container instance we can ensure that it will be
"Running" when we execute our command to fetch the IP address.