Fix sporadically failing payroll runs feature spec
There is no connection between the ordering of PayrollRun#claims and
PayrollRun#payments, so if we wish to refer to the same payment multiple times
inside a spec then we need to consistently refer to only one of these
relations.

I ran the test suite ten times in a row without any failures, so
hopefully this is fixed.