Allow unverified addresses from GOV.UK Verify
We're getting a few support queries from people saying we have the
wrong address for them. We don't need this value to be verified as we
just forward it to the payroll provider and HMRC.

We may want to rename the `verfied_attributes` field as the address is
no longer verfied.