Remove role="main" from main element
As we are using a main element to denote the dominant content of the
body we do not need to specify its role. A main element with a role of
main would return a warning when put through a validator.

If we didn't have a main then it would be appropriate to add this to the
the parent div.