Change links that look like buttons to be buttons
We have a couple of links that are styled to look like buttons and have
the appropriate aria role so that screen readers will put them up
however, they do not function as a keyboard only user wold expect them
to. Users were unable to give the buttons focus and press space to
trigger them.

This PR swaps the links to be actual buttons so they will function as
expected.