Use a standard job for recording claim events
Now we've refactored the GeckoboardEvent class, it makes less sense to
have seperate jobs for when a claim has been submitted and when a claim
has been paid, so we can unite them into one job, which accepts an array of
claim IDs, the event type, and the date_field to use for the submitted_at field.
This can also be reused for the backfilling.