Fix flickering specs when testing Geckoboard data
Before, we would have ocassional test flicker when testing that arrays of data
had been sent to Geckoboard. This adds a support method which sorts the
expected payload and the actual payload by reference, so we can make sure
they both match up, regardless of the order they are sent. It also has the
added benefit of DRYing up the tests as all the marshalling is done in one
place.

# Conflicts:
#	spec/jobs/record_claim_event_job_spec.rb