Override db:setup to seed data_migrations table
Ensures that running db:setup will result in a database where any
existing data migrations will be treate as having already been run.

When a new database is set up we want to make sure all existing data
migrations are listed in the data_migrations table, otherwise they will
be treated as having not been run and will be executed the next time
db:migrate:with_data is called. This causes issues as some of the older
data migrations will no longer execute because the schema has moved on
since then.

By overriding the db:setup we make sure that a developer that runs
that (as well as db:reset) will end up with a clean database where all
migrations and data migrations are marked as having been executed.