Nightly job to purge old unsubmitted claims
This will run every day at midnight and remove any unsubmitted claims
that are more than 24 hours old so we don't hang on to user data that we
don't need and can't use.

Because this inherits from `CronJob` it will automatically get picked
up and scheduled by `CronJobScheduler` when the `jobs:schedule` rake
task is run.