Make sure datasets are uniqued by reference
When updating Geckoboard, we need to make sure that any additional or
updated data about a claim can be submitted against the same claim
reference, rather than creating a new one, if it exists. Setting
`unique_by` means that Geckoboard treats any subsequent push of data
with a given reference as an update action, rather than a create action.