Move Geckoboard to a single dataset
This moves the Geckoboard data model from using one dataset per event
type to having one dataset for all submitted claims regardless of
status.

Where a claim has not been paid the date and time of `paid_at` is set to
1970-01-01. This is to work around Geckoboard not supporting null dates.
The date is visibly wrong where incorrectly included in visualisations,
and claims which should not be included can be filtered using the `paid`
field.

This also removes the now defunct data migration for bulk filling
Geckoboard. Not only would this migration fail to run, but there is no
meaningful way to make it work with the current data model.

Co-authored-by: Stuart Harrison <stuart@dxw.com>