Change existing session to use radios
This PR builds upon the existing interruption card that was used with
some usability improvements. We now ask the user if they want to abandon
their current claim and use radio buttons for them to select their
response.

We have also tided up some of the CSS that is no longer needed.