Add claim check data to Geckoboard
The check dataset on Geckoboard now includes if a claim has been checked
or not, and if so the date and time and result of the check. Where a
claim has not been checked the timestamp is set to 1970-01-01 to work
around the lack of NULL datetime fields in Geckoboard.

Co-authored-by: Stuart Harrison <stuart@dxw.com>