Disable alerts by default in the test environment
Most of the alerts are intended to let us know that the application
isn't working as expected. For the test environment we mainly care
about if the infrastructure has deployed or not. eg. Alerts letting us
know that the CPU usage is high during a deployment are just noise.

I've left the service health alerts in there as they only happen when
there's a problem with the infrastructure.