Styling fixes on interrpution card
Prior to this the button was changing to green whilst it was disabled
and the next page was loading. This keeps the button correctly coloured
and uses the standard button disabled opacity to indicate that it cannot
be clicked on multiple times.

It also changes the background colour on the interrpution card.
Previously we were using a default gov blue colour but as this is in the
community backlog then it makes sense to take the colour Verify use
and tweak it to provide AAA contrast.