Allow soft-deleted KeyVaults to be recovered with a flag
We considered making this automatic with another step of the script
that checks to see if a soft-deleted KeyVault exists and change the
`createMode` depending on the result.

However we decided against it. Deleting a KeyVault is quite a
purposeful action, the expected result from the next deployment should
error loudly to the user to let them know they need to either delete
the KeyVault permanently or recover it explicitly.

The KeyVault can be recovered using the `--recover-keyvault` flag when
calling the azure-deploy script.