Add named log tag for environment
This allows the various deployment environments to be distinguished in
the logs, since the `ENVIRONMENT` variable is consistent across both
development and production.

`config.log_tags` is a Ruby built-in which can be overridden as part of
Semantic Logger (see
https://rocketjob.github.io/semantic_logger/rails.html#named-tags) to
add additional parameters to all requests.

ENVIRONMENT_NAME is now also set to `local` in `.env.development`, so
developers working locally will have the variable set by default.